Processing Scripts
Updated and created processing scripts that give a good outline of the
structure. The big issue will be the spatial join scripts. My current
idea is to use a template in the SurveyTrees gdb and to use that for
field names. If that doesn't work I'll have to figure out how to hard
code batch spatial joining.